[otbn,rtl] Prevent LSU error propagation during a delayed escalation
If a e.g. predec error happens during the first cycle of a LSU
instruction, due to the registering of the error the LSU will still
return a result but its error signal is X. We prevent the propagation of
the X signal into the error bits by ignoring the LSU error if there is
already an escalation ongoing.

Signed-off-by: Pascal Etterli <pascal.etterli@lowrisc.org>